requirements.txt
# GameAgent: unified requirements
# Use: pip install -r requirements.txt

torch>=2.1.0
transformers>=4.45.0,<5.0.0
accelerate>=1.0.0
datasets>=3.0.0
numpy>=1.26.0
scipy>=1.14.0
matplotlib>=3.9.0
tqdm>=4.66.0
pandas>=2.2.0
pyyaml>=6.0.1
huggingface_hub>=0.25.0
evaluate>=0.4.0

# Training frameworks (pinned for NashDPOTrainer API compatibility)
trl>=0.15.0,<0.17.0
peft>=0.15.0

# Robust reward models
sentence-transformers>=3.0.0
